backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-adk>=0.3.0
google-genai>=1.24.0
google-cloud-firestore>=2.21.0
httpx>=0.28.0
websockets>=12.0
python-dotenv>=1.1.1
